+++ insult_error.py
@@ -3,7 +3,19 @@ import random
 from collections import namedtuple
 
 
+# the main event! ----------
+
+
 class InsultError(Exception):
+    """
+    Randomized insulting error name and message
+
+    Arguments:
+        *args: positional arguments for anything you would normally pass when creating an
+            Exception object, usually just a message string
+        rating: optionally, set a limit on how offensive the error can be as a number between
+            1 and 10 (1 being tamest, and 10 being meanest), defaults to 5
+    """
 
     RATING = 5
 
@@ -11,63 +23,89 @@ class InsultError(Exception):
         if rating is None:
             rating = self.RATING
         # hack the error name to be a random insult rated <= rating
-        name = random.choice([x.name for x in names if x.rating <= rating])
+        name = random.choice([x.name for x in insult_names if x.rating <= rating])
         self.__class__.__name__ = name
+        # hack the error module name so it pretends to be a built-in, yields a prettier traceback
+        # see: traceback.format_exception_only method
+        self.__class__.__module__ = 'builtins'
         # if no message is provided, select a random insult
         if not args:
-            msg = random.choice([x.msg for x in messages if x.rating <= rating])
+            msg = random.choice([x.msg for x in insult_messages if x.rating <= rating])
             args = (msg,)
         self.args = args
 
 
-def fuck_you(type, value, traceback):
-    """
-    sys.excepthook standin that injects an error but leaves the traceback
-    """
-    insult_error = InsultError()
-    # Call the original excepthook so we get a traceback printed out
-    sys.__excepthook__(type(insult_error), insult_error, traceback)
+# automatic insults ----------
 
-sys.excepthook = fuck_you
+
+def insulthook(rating=None, preserve_msg=False):
+    """
+    Return a sys.excepthook standin that injects InsultError, but leaves the traceback
+
+    Arguments:
+        rating: limit for how offensive you want the error and messages to be
+            (1 being tamest and 10 being meanest)
+        preserve_msg: set True to keep the message string from the original error,
+            or False to replace it with an insult
+    """
+    def f(type, value, traceback):
+        """sys.excepthook standin that injects an error but leaves the traceback"""
+        if str(value) and preserve_msg:
+            exc = InsultError(str(value), rating=rating)
+        else:
+            exc = InsultError(rating=rating)
+        sys.__excepthook__(type(exc), exc, traceback)
+    return f
+
+
+def always_insult_me(rating=None, preserve_msg=False):
+    """Replace uncaught exceptions with insults at or below rating"""
+    sys.excepthook = insulthook(rating, preserve_msg)
+
+
+def dont_always_insult_me():
+    """Stop replacing uncaught exceptions with insults"""
+    sys.excepthook = sys.__excepthook__
+
 
 # insulting error names ----------
 
 
-Name = namedtuple('Name', ('name', 'rating'))
+_Name = namedtuple('Name', ('name', 'rating'))
 
-names = [
-    Name(rating=5, name='FuckYouBuddy'),
-    Name(rating=5, name='FuckYouFriend'),
-    Name(rating=4, name='FrickYouPal'),
-    Name(rating=1, name='NotThisAgain'),
-    Name(rating=1, name='ForGodsSake'),
-    Name(rating=1, name='AreYouSerious'),
+insult_names = [
+    _Name(rating=5, name='FuckYouBuddy'),
+    _Name(rating=5, name='FuckYouFriend'),
+    _Name(rating=4, name='FrickYouPal'),
+    _Name(rating=1, name='NotThisAgain'),
+    _Name(rating=1, name='ForGodsSake'),
+    _Name(rating=1, name='AreYouSerious'),
 ]
 
 
 # insulting error messages ----------
 
 
-Message = namedtuple('Message', ('msg', 'rating'))
+_Message = namedtuple('Message', ('msg', 'rating'))
 
-messages = [
-    Message(rating=1, msg="Your program is bad and you should feel bad"),
-    Message(rating=1, msg="I envy people who have never met you"),
-    Message(rating=1, msg="You're impossible to underestimate"),
-    Message(rating=1, msg="I don't have the time or the crayons to explain this to you"),
-    Message(rating=1, msg="Don't believe everything you think"),
-    Message(rating=1, msg="I hope this isn't your day job"),
-    Message(rating=1, msg="You're killing your mother right now"),
-    Message(rating=1, msg="I hear that fast food place is still hiring"),
-    Message(rating=1, msg="If you were on fire and I had water, I'd drink it."),
-    Message(rating=1, msg="You messed up - don't try again for your own good"),
-    Message(rating=1, msg=(
+insult_messages = [
+    _Message(rating=1, msg="Your program is bad and you should feel bad"),
+    _Message(rating=1, msg="I envy people who have never met you"),
+    _Message(rating=1, msg="You're impossible to underestimate"),
+    _Message(rating=1, msg="I don't have the time or the crayons to explain this to you"),
+    _Message(rating=1, msg="Don't believe everything you think"),
+    _Message(rating=1, msg="I hope this isn't your day job"),
+    _Message(rating=1, msg="You're killing your mother right now"),
+    _Message(rating=1, msg="I hear that fast food place is still hiring"),
+    _Message(rating=1, msg="If you were on fire and I had water, I'd drink it."),
+    _Message(rating=1, msg="You messed up - don't try again for your own good"),
+    _Message(rating=1, msg=(
         "This is like being in a house built by a child using nothing but a "
         "hatchet and a picture of a house [xkcd.com/1513]")),
-    Message(rating=1, msg=(
+    _Message(rating=1, msg=(
         "It's like a salad recipe written by a corporate lawyer using a phone "
         "autocorrect that only knew excel formulas [xkcd.com/1513]")),
-    Message(rating=1, msg=(
+    _Message(rating=1, msg=(
         "It's like someone took a transcript of a couple arguing at Ikea and "
         "made random edits until it compiled without errors [xkcd.com/1513]")),
     # Message(rating=1, msg=(
@@ -94,7 +132,7 @@ messages = [
     #     "Your code looks like you read Turing's 1936 paper on computing and a "
     #     "page of Javascipt example code and guessed at everything in between "
     #     "[xkcd.com/1833]")), 
-    Message(rating=1, msg=(
+    _Message(rating=1, msg=(
         "It's like a leet-speak translation of a manifesto by a survivalist"
         "cult leader who's for some reason obsessed with memory allocation"
         "[xkcd.com/1833]")),